refactor(clients): tidy imports
